Refactoring Slider
