+++ asset-manifest.json
@@ -2,8 +2,8 @@
   "files": {
     "static/js/1.cbe5ed3b.chunk.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/1.cbe5ed3b.chunk.js",
     "static/js/1.cbe5ed3b.chunk.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/1.cbe5ed3b.chunk.js.map",
-    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.9c14a0e8.chunk.js",
-    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.9c14a0e8.chunk.js.map",
+    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.54d8ee3d.chunk.js",
+    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.54d8ee3d.chunk.js.map",
     "runtime~main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js",
     "runtime~main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js.map",
     "index.php": "/react-wp/wp-content/themes/inu-v2-b/index.php"
