+++ react-src/src/components/Slides/Slides.md
@@ -1,0 +1,1 @@
+## Slides
