+++ react-src/src/components/Slides/index.js
@@ -1,0 +1,2 @@
+export { default } from "./Slides";
+export { SlidesPropTypes, SlidesDefaultProps } from "./Slides";
